requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.10.1
PyPDF2==3.0.1
requests==2.31.0
gunicorn==21.2.0

